# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Pleco API
  version: 1.0.0
  description: |
    OpenAPI reference for pleco-api.

    Most API responses use a standard envelope:
    - success: `status`, `message`, optional `data`, optional `meta`
    - error: `status`, `message`, optional `errors`
servers:
  - url: http://localhost:8080
    description: Local development
  - url: http://localhost
    description: Docker Compose gateway
  - url: https://pleco-api.onrender.com
    description: Production
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: User registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeMessage'
        '400':
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeValidationError'
  /auth/login:
    post:
      tags: [Auth]
      summary: Login with email and password
      description: Creates a refresh-token session and lets the API assign or reuse the `pleco_device_id` HttpOnly cookie.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login success
          headers:
            Set-Cookie:
              description: HttpOnly cookies named `pleco_refresh_token` and `pleco_device_id`
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeAccessToken'
        '401':
          description: Invalid credentials or unverified email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    RegisterRequest:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
...
    AccessToken:
      type: object
      properties:
        access_token:
          type: string
...
    FieldError:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
    EnvelopeMessage:
      type: object
      properties:
        status:
          type: string
          example: success
        message:
          type: string
    EnvelopeError:
      type: object
      properties:
        status:
          type: string
          example: error
        message:
          type: string
    EnvelopeValidationError:
      type: object
      properties:
        status:
          type: string
          example: error
        message:
          type: string
          example: Validation failed
        errors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
...
    EnvelopeAccessToken:
      type: object
      properties:
        status:
          type: string
          example: success
        message:
          type: string
        data:
          $ref: '#/components/schemas/AccessToken'
...